Move tests to modules they test (loader.rs and extraction.rs)
Reorganized test structure for better locality:

Before:
- lexplore.rs: 451 lines (35 module code + 416 lines of tests)
- loader.rs: 805 lines (code only)
- extraction.rs: 107 lines (code only)

After:
- lexplore.rs: 35 lines (module declarations and re-exports only)
- loader.rs: 1171 lines (805 code + 366 lines of loader tests)
- extraction.rs: 130 lines (107 code + 23 lines of extraction tests)

Benefits:
- Tests are co-located with the code they test
- Easier to find relevant tests when modifying code
- Clear separation: loader tests test loading/parsing, extraction tests test element extraction
- All 45 tests still passing, maintained 100% coverage
- Follows Rust best practices for module-level testing

Test distribution:
- 43 loader tests (file loading, parsing, tokenizing, fluent API)
- 2 extraction tests (AST element extraction helpers)
